Fixing various bugs related to type conversions. Some minor pedantic warnings with clangd compiler still persist.

+++ src/hash_table.c
@@ -68,10 +68,10 @@ void ht_del_hash_table(ht_hash_table *ht) {
 // Returns tha hash of 's', an int between 0 and 'm'
 static unsigned int ht_hash(const char *s, const int a, const int m) {
   long hash = 0;
-  const int len_s = strlen(s);
+  const unsigned int len_s = strlen(s);
 
   for (size_t i = 0; i < len_s; i++) {
-    hash += (long)pow(a, len_s - (i + 1)) * s[i];
+    hash += (long)pow(a, (double)len_s - (double)(i + 1)) * s[i];
     hash = hash % m;
   }
 
@@ -80,10 +80,10 @@ static unsigned int ht_hash(const char *s, const int a, const int m) {
 
 static int ht_get_hash(const char *s, const int num_buckets,
                        const int attempt) {
-  const int hash_a = ht_hash(s, HT_PRIME_1, num_buckets);
-  const int hash_b = ht_hash(s, HT_PRIME_2, num_buckets);
+  const unsigned int hash_a = ht_hash(s, HT_PRIME_1, num_buckets);
+  const unsigned int hash_b = ht_hash(s, HT_PRIME_2, num_buckets);
 
-  return (hash_a + (attempt * (hash_b + 1))) % num_buckets;
+  return (int)(hash_a + (attempt * (hash_b + 1))) % num_buckets;
 }
 
 void ht_insert(ht_hash_table *ht, const char *key, const char *value) {
